refactor(dashboard): split peg loading skeleton
